Add prepaid flights ⚠️ breaks DB ⚠️
As only not yet deployed models of bookkeeping were affected, I did not
bother making proper migrations, when fixing the paid-typo ...
Thus you should delete the development database db.sqlite3 before
`$ python manage.py migrate`.

This also made it cheap to add `report` to `Purchase`, consistent with
`Run`, `Bill`, and `Expense`. While a link to a `Report` would only be
necessary for `Expense`, it is convenient for the others (e.g. in
`Report.revenue`) and makes sure, the data stay consistent.